# expression: implement vectorized evaluation for 'builtinCoalesceJSONSig'