@qw4990 2019-10-25
All functions in this document are suggested to be vectorized by using `go-gen`, and @gtygo is working on using `go-gen` to generate all Coalesce functions.
So how about waiting for @gtygo 's PR?